impl(testing): Phase 1B reshape — _shared/ + toxics/ + driver/ + 3 mergers (#541)
Closes #541. Implements Phase 1B per architect plan r2 (#541 comment 4411296014).

What changed
------------
Reorganizes 8 testing-infra files at `packages/protocol/src/testing/` root
into `conformance/_shared/` (cross-layer helpers used by every property)
and `toxics/` (subsystem-local). External barrel surface
`@moltzap/protocol/testing` is byte-identical pre/post; only relative
import paths inside the barrel shift.

Layout
------
- `conformance/_shared/driver/{test-client,test-server,index}.ts` — wire-driver harness
- `conformance/_shared/frame-mutator.ts` — renamed from `codec.ts`; inlines `FrameSchemaError`
- `conformance/_shared/captures.ts` — moved
- `conformance/_shared/canonicalize.ts` — moved
- `conformance/_shared/errors.ts` — TransportClosed/Io, RpcTimeout/Response, RealServerAcquire, TestingError union
- `conformance/_shared/test-fixtures.ts` — merges branded-ids + agent-registration
- `toxics/errors.ts` — `ToxicControlError` (subsystem-local)
- 8 deletions at testing/ root + 38 consumer-import rewrites + 1 outside-testing redirect
  (`network/actor-model.test.ts`: branded-ids → barrel)

Verifications
-------------
- `pnpm --filter @moltzap/protocol build` — green
- `pnpm --filter @moltzap/protocol test` — 114 pass, 1 todo (11 files)
- `pnpm build` workspace-wide — green
- `pnpm --filter @moltzap/server-core test:conformance` — passed=44 deferred=5 unavailable=1 failed=0
- `bash packages/protocol/scripts/check-divergence-proofs.sh` — OK (26 registrars proof-required, 38 legacy exemptions)
- `tsc --emitDeclarationOnly` barrel diff — only path metadata changed; every named symbol identical
- Scripts audit — no scripts/** references the moved files
- Docs audit — 4 historical refs in docs/plans/phase-12-jsonrpc-finalization-2026-05.md
  left per architect §5 NOT-in-scope and §9 #6 mitigation (historical plan doc)
