+++ packages/protocol/src/testing/conformance/_shared/canonicalize.ts
