+++ packages/protocol/src/testing/conformance/client/adversity.ts
@@ -32,7 +32,7 @@ import {
   agentId,
   conversationId as toConversationId,
   messageId,
-} from "../../branded-ids.js";
+} from "../_shared/test-fixtures.js";
 import type { ClientConformanceRunContext } from "./runner.js";
 import { PropertyUnavailable, registerProperty } from "../_shared/registry.js";
 import {
