+++ packages/protocol/src/testing/conformance/client/rpc-semantics.ts
@@ -18,7 +18,7 @@ import { responseFrame } from "../../../transport/wire.js";
 import type { ClientConformanceRunContext } from "./runner.js";
 import { registerProperty } from "../_shared/registry.js";
 import { acquireFixture, invariant } from "./_fixtures.js";
-import { isRequestFrame } from "../../codec.js";
+import { isRequestFrame } from "../_shared/frame-mutator.js";
 
 import { AgentsList } from "../../../identity/methods.js";
