+++ packages/protocol/src/testing/conformance/identity/authority-positive.ts
@@ -6,8 +6,8 @@
  */
 import { Effect } from "effect";
 import { ConversationsList } from "../../../task/methods.js";
-import { makeTestClient } from "../../test-client.js";
-import { registerTestAgent } from "../../agent-registration.js";
+import { makeTestClient } from "../_shared/driver/test-client.js";
+import { registerTestAgent } from "../_shared/test-fixtures.js";
 import type { ConformanceRunContext } from "../_shared/runner.js";
 import {
   PropertyInvariantViolation,
