+++ packages/protocol/src/testing/conformance/task/model-equivalence.ts
@@ -38,8 +38,8 @@ import {
 } from "../../arbitraries/rpc.js";
 import { applyCall } from "../../models/dispatch.js";
 import { initialReferenceState } from "../../models/state.js";
-import { makeTestClient } from "../../test-client.js";
-import { registerTestAgent } from "../../agent-registration.js";
+import { makeTestClient } from "../_shared/driver/test-client.js";
+import { registerTestAgent } from "../_shared/test-fixtures.js";
 import type { ConformanceRunContext } from "../_shared/runner.js";
 import {
   PropertyInvariantViolation,
