+++ packages/protocol/src/testing/conformance/transport/adversity-latency-resilience.ts
@@ -5,7 +5,7 @@
  */
 import { Effect } from "effect";
 import { defaultToxicProfile } from "../../toxics/defaults.js";
-import { isNotificationFrame } from "../../codec.js";
+import { isNotificationFrame } from "../_shared/frame-mutator.js";
 import { MessagesSend } from "../../../task/methods.js";
 import type { ConformanceRunContext } from "../_shared/runner.js";
 import {
