+++ packages/protocol/src/testing/conformance/transport/adversity-reset-peer-recovery.ts
@@ -8,7 +8,7 @@
  */
 import { Clock, Effect, Either } from "effect";
 import { defaultToxicProfile } from "../../toxics/defaults.js";
-import { TransportClosedError } from "../../errors.js";
+import { TransportClosedError } from "../_shared/errors.js";
 import { ConversationsList } from "../../../task/methods.js";
 import type { ConformanceRunContext } from "../_shared/runner.js";
 import { PropertyUnavailable } from "../_shared/registry.js";
