+++ packages/protocol/src/testing/conformance/transport/caller-controlled-app-callback-timeout.ts
@@ -13,8 +13,8 @@
  */
 import { Effect, Either } from "effect";
 import { DispatchAuthorize } from "../../../app/methods.js";
-import { makeTestClient } from "../../test-client.js";
-import { registerTestAgent } from "../../agent-registration.js";
+import { makeTestClient } from "../_shared/driver/test-client.js";
+import { registerTestAgent } from "../_shared/test-fixtures.js";
 import type { ConformanceRunContext } from "../_shared/runner.js";
 import {
   PropertyInvariantViolation,
