+++ packages/protocol/src/testing/conformance/transport/malformed-frame-handling.ts
@@ -7,8 +7,8 @@ import * as fc from "fast-check";
 import { Effect, Either } from "effect";
 import { arbitraryMalformedFrame } from "../../arbitraries/frames.js";
 import { AgentsList } from "../../../identity/methods.js";
-import { makeTestClient } from "../../test-client.js";
-import { registerTestAgent } from "../../agent-registration.js";
+import { makeTestClient } from "../_shared/driver/test-client.js";
+import { registerTestAgent } from "../_shared/test-fixtures.js";
 import type { ConformanceRunContext } from "../_shared/runner.js";
 import { assertProperty, registerProperty } from "../_shared/registry.js";
