+++ packages/protocol/src/testing/conformance/transport/request-id-uniqueness.ts
@@ -6,9 +6,9 @@ import * as fc from "fast-check";
 import { Effect } from "effect";
 import { ConversationsList } from "../../../task/methods.js";
 import { type JsonRpcId } from "../../../transport/wire.js";
-import { isRequestFrame, isResponseFrame } from "../../codec.js";
-import { makeTestClient } from "../../test-client.js";
-import { registerTestAgent } from "../../agent-registration.js";
+import { isRequestFrame, isResponseFrame } from "../_shared/frame-mutator.js";
+import { makeTestClient } from "../_shared/driver/test-client.js";
+import { registerTestAgent } from "../_shared/test-fixtures.js";
 import type { ConformanceRunContext } from "../_shared/runner.js";
 import { assertProperty, registerProperty } from "../_shared/registry.js";
