+++ packages/protocol/src/testing/conformance/transport/request-well-formedness.ts
@@ -15,9 +15,9 @@ import {
   ResponseFrameSchema,
   type ResponseFrame,
 } from "../../../transport/wire.js";
-import { isRequestFrame, isResponseFrame } from "../../codec.js";
-import { makeTestClient } from "../../test-client.js";
-import { registerTestAgent } from "../../agent-registration.js";
+import { isRequestFrame, isResponseFrame } from "../_shared/frame-mutator.js";
+import { makeTestClient } from "../_shared/driver/test-client.js";
+import { registerTestAgent } from "../_shared/test-fixtures.js";
 import type { ConformanceRunContext } from "../_shared/runner.js";
 import { assertProperty, registerProperty } from "../_shared/registry.js";
