+++ packages/protocol/src/testing/conformance/transport/rpc-map-coverage.ts
@@ -8,9 +8,9 @@
 import * as fc from "fast-check";
 import { Effect } from "effect";
 import { arbitraryCallFor } from "../../arbitraries/rpc.js";
-import { isRequestFrame, isResponseFrame } from "../../codec.js";
-import { makeTestClient } from "../../test-client.js";
-import { registerTestAgent } from "../../agent-registration.js";
+import { isRequestFrame, isResponseFrame } from "../_shared/frame-mutator.js";
+import { makeTestClient } from "../_shared/driver/test-client.js";
+import { registerTestAgent } from "../_shared/test-fixtures.js";
 import type { ConformanceRunContext } from "../_shared/runner.js";
 import {
   PropertyInvariantViolation,
