+++ packages/protocol/src/testing/conformance/transport/schema-exhaustive-fuzz.ts
@@ -13,8 +13,8 @@ import * as fc from "fast-check";
 import { Effect } from "effect";
 import { allRpcMethods, arbitraryCallFor } from "../../arbitraries/rpc.js";
 import { AgentsList } from "../../../identity/methods.js";
-import { makeTestClient } from "../../test-client.js";
-import { registerTestAgent } from "../../agent-registration.js";
+import { makeTestClient } from "../_shared/driver/test-client.js";
+import { registerTestAgent } from "../_shared/test-fixtures.js";
 import type { ConformanceRunContext } from "../_shared/runner.js";
 import {
   PropertyInvariantViolation,
